Organize server configurations
