+++ src/constants/index.ts
@@ -3,3 +3,12 @@ export const API_PROCESS_PATH = 'process';
 export const API_POINTS_PATH = 'points';
 
 export const API_VERSION_ONE_PREFIX = 'v1';
+
+export const LOG_LEVEL = {
+  TRACE: 'trace',
+  DEBUG: 'debug',
+  INFO: 'info',
+  WARN: 'warn',
+  ERROR: 'error',
+  FATAL: 'fatal',
+};
