+++ src/infrastructure/config/fastify-config.ts
@@ -1,0 +1,8 @@
+import Fastify from 'fastify';
+
+export const createFastifyInstance = () => {
+  return Fastify({
+    logger: true,
+    ignoreTrailingSlash: true,
+  });
+};
