+++ src/interface/controllers/get-receipt-points/get-receipt-points-controller.ts
