+++ src/interface/controllers/process-receipt/process-receipt-controller.test.ts
@@ -1,9 +1,9 @@
+import { FastifyReply, FastifyRequest } from 'fastify';
 import { getReceiptService } from '../../../application/core/domain/services/receipt-service-factory';
 import { HTTPError } from '../../errors/http-error';
-import { GetPointsResponse } from '../../../types/http/get-receipt-points';
-import { getReceiptPointsController } from './get-receipt-points';
+import { Receipt, mockReceipt } from '../../../types/domain/receipt';
+import { processReceiptController } from './process-receipt-controller';
 import { IReceiptService } from '../../../application/core/domain/services/receipt-service-interface';
-import { FastifyRequest, FastifyReply } from 'fastify';
 
 jest.mock(
   '../../../application/core/domain/services/receipt-service-factory',
@@ -12,9 +12,9 @@ jest.mock(
   })
 );
 
-describe('getReceiptPointsController', () => {
+describe('processReceiptController', () => {
   let mockReply: FastifyReply;
-  let mockRequest: FastifyRequest<{ Params: { id: string } }>;
+  let mockRequest: FastifyRequest<{ Body: Receipt }>;
   let mockedReceiptService: jest.Mocked<IReceiptService>;
 
   beforeEach(() => {
@@ -26,11 +26,11 @@ describe('getReceiptPointsController', () => {
     } as unknown as FastifyReply;
 
     mockRequest = {
-      params: { id: 'receipt1' },
-    } as unknown as FastifyRequest<{ Params: { id: string } }>;
+      body: mockReceipt,
+    } as FastifyRequest<{ Body: Receipt }>;
 
     mockedReceiptService = {
-      processReceipt: jest.fn(),
+      processReceipt: jest.fn().mockReturnValue('12345'), // Mock processReceipt to return a receipt ID
       findReceiptPoints: jest.fn(),
     };
 
@@ -38,35 +38,32 @@ describe('getReceiptPointsController', () => {
     (getReceiptService as jest.Mock).mockReturnValue(mockedReceiptService);
   });
 
-  it('should return 200 and the receipt points when the receipt is found', async () => {
-    const mockResponse: GetPointsResponse = { points: 10 };
-    mockedReceiptService.findReceiptPoints.mockReturnValue(mockResponse);
+  it('should return 201 and the receipt ID when the receipt is successfully processed', async () => {
+    processReceiptController(mockRequest, mockReply);
 
-    await getReceiptPointsController(mockRequest, mockReply);
-
-    expect(mockReply.status).toHaveBeenCalledWith(200);
-    expect(mockReply.send).toHaveBeenCalledWith(mockResponse);
+    expect(mockReply.status).toHaveBeenCalledWith(201);
+    expect(mockReply.send).toHaveBeenCalledWith({ id: '12345' });
   });
 
-  it('should return 404 and an error message when receipt is not found', async () => {
-    const errorMessage = 'Receipt not found';
-    mockedReceiptService.findReceiptPoints.mockImplementation(() => {
-      throw new HTTPError(errorMessage, 404);
+  it('should return 400 and an error message when HTTPError is thrown', async () => {
+    const errorMessage = 'Invalid receipt data';
+    mockedReceiptService.processReceipt.mockImplementation(() => {
+      throw new HTTPError(errorMessage, 400);
     });
 
-    await getReceiptPointsController(mockRequest, mockReply);
+    processReceiptController(mockRequest, mockReply);
 
-    expect(mockReply.status).toHaveBeenCalledWith(404);
+    expect(mockReply.status).toHaveBeenCalledWith(400);
     expect(mockReply.send).toHaveBeenCalledWith({ error: errorMessage });
   });
 
   it('should return 500 and a generic error message for unexpected errors', async () => {
     const errorMessage = 'Some unexpected error';
-    mockedReceiptService.findReceiptPoints.mockImplementation(() => {
+    mockedReceiptService.processReceipt.mockImplementation(() => {
       throw new Error(errorMessage);
     });
 
-    await getReceiptPointsController(mockRequest, mockReply);
+    processReceiptController(mockRequest, mockReply);
 
     expect(mockReply.status).toHaveBeenCalledWith(500);
     expect(mockReply.send).toHaveBeenCalledWith({
