+++ src/server.ts
@@ -1,36 +1,21 @@
-import Fastify from 'fastify';
-import processReceiptRoute from './interface/routes/v1/receipts/process/route';
-import getPointsRoute from './interface/routes/v1/receipts/:id/points/route';
-import { API_VERSION_ONE_PREFIX } from './constants';
+import { registerRoutes } from './interface/routes';
+import { createFastifyInstance } from './infrastructure/config/fastify-config';
+import { startServer } from './infrastructure/config/server-config';
+import { versionRedirectHook } from './interface/hooks/version-redirect-hook';
 
 // Instantiate Fastify
-const fastify = Fastify({
-  logger: true,
-  ignoreTrailingSlash: true,
-});
+const fastify = createFastifyInstance();
 
-// Basic route to check if Fastify is working
+// Basic health check route
 fastify.get('/', async () => {
   return { message: 'Receipt Processor API' };
 });
 
 // Register routes
-fastify.register(processReceiptRoute, {
-  prefix: `/${API_VERSION_ONE_PREFIX}/`,
-});
-fastify.register(getPointsRoute, {
-  prefix: `/${API_VERSION_ONE_PREFIX}`,
-});
+registerRoutes(fastify);
+
+// Register hooks
+fastify.addHook('onRequest', versionRedirectHook);
 
 // Start the server
-const start = async () => {
-  fastify.listen({ port: 3000, host: '0.0.0.0' }, function (err, address) {
-    if (err) {
-      fastify.log.error(err);
-      process.exit(1);
-    }
-    console.log(`Server listening on: ${address}`);
-  });
-};
-
-start();
+startServer(fastify);
